sort vault servers naturally
