+++ lib/samson/natural_order.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+module Samson
+  module NaturalOrder
+    # split words into string and numbers to better sort them "a1b2" -> ["a", 1, "b", 2]
+    # this avoids pod2 being sorted after pod10
+    def self.covert(name)
+      name.split(/(\d+)/).each_with_index.map { |x, i| i.odd? ? x.to_i : x }
+    end
+  end
+end
